mantis leifos ticket 1196

+++ classes/class.ilPermissionManagerAction.php
@@ -324,7 +324,7 @@ class ilPermissionManagerAction
 
     public function getTimingVisibility() : int
     {
-        return ($this->timing_visibility ?? 0);
+        return (int) ($this->timing_visibility ?? 0);
     }
 
     public function setTimingVisibility(int $a_stat)
